fix: cancel TSQ goroutine once exec command completes
In the case of SSH ControlMaster connections with a PTY, the SSH
connection can stay up without an active SSH channel (i.e. after the
command execution completes). If this happens the ssh connection context
will not be cancelled, as the connection is still up.

Wrapping the raw SSH context and deferring the cancel of it before
passing it to exec.StreamWithContext() and newTermSizeQueue() ensures
that once command execution completes, the context is cancelled and the
TSQ goroutine will exit immediately.

If the context is not cancelled after command execution completes, the
TSQ goroutine will spin trying to write to a closed winch channel and
spew EOF errors into stderr.