feat: update last_used timestamp for SSH keys
Update the last_used timestamp to the current time whenever an SSH key
is used to either generate a token or query exec permissions on a Lagoon
environment.

The timestamp is updated every time the key is used, regardless of
whether or not the permission check or token generation succeeds.

Timestamps are converted and stored in UTC.