feat: rename existing metric to match all the others
This was originally prefixed sshportalapi because it was only observed
by the ssh-portal-api service. However we now have HistogramVecs which
are observed by multiple services so it makes sense to just use the
sshportal prefix instead.

This is a breaking change, but only of metrics. We are still v0.x so
users will just have to update their dashboards.